refactor controls
