+++ src/pages/levelCreator/Controls.tsx
@@ -1,29 +1,9 @@
-import {
-  Casino as CasinoIcon,
-  CloudDownload as CloudDownloadIcon,
-  Description as DescriptionIcon,
-  EditRoad as EditRoadIcon,
-  ExitToApp as ExitToAppIcon,
-  Extension as ExtensionIcon,
-  Help as HelpIcon,
-  Home as HomeIcon,
-  Lightbulb as LightbulbIcon,
-  LocalShipping as LocalShippingIcon,
-  Park as ParkIcon,
-  People as PeopleIcon,
-  Save as SaveIcon,
-  Traffic as TrafficIcon,
-} from "@mui/icons-material"
 import { type FC, useEffect, useState } from "react"
 import { Divider } from "@mui/material"
 import type Phaser from "phaser"
 
-import * as miniDrawers from "../../components/miniDrawers"
-import * as tilesets from "../../phaser/tilesets"
-import EndpointsControls from "./endpoints/Controls"
-import ObstaclesControls from "./obstacles/Controls"
-import RoadControls from "./road/Controls"
-import SceneryControls from "./scenery/Controls"
+import * as items from "./items"
+import { MiniDrawer } from "../../components/miniDrawers"
 import { ZoomControls } from "../../phaser"
 import { usePhaserGameContext } from "../../app/hooks"
 
@@ -36,39 +16,12 @@ const Controls: FC<ControlsProps> = () => {
     activeSceneKeys,
   } = usePhaserGameContext()
 
-  // States for each box's tool - persist the last selected tool for each box.
-  const [roadTool, setRoadTool] =
-    useState<Phaser.Types.Scenes.Create.Toolbox.Road["tool"]>("add")
-  const [endpointsTool, setEndpointsTool] = useState<
-    Phaser.Types.Scenes.Create.Toolbox.Endpoints["tool"]
-  >(tilesets.IDs.Endpoints.CFC.Warehouse.DEFAULT)
-  const [sceneryTool, setSceneryTool] = useState<
-    Phaser.Types.Scenes.Create.Toolbox.Scenery["tool"]
-  >(tilesets.IDs.Scenery.Nature.BUSH)
-  const [obstaclesTool, setObstaclesTool] = useState<
-    Phaser.Types.Scenes.Create.Toolbox.Obstacles["tool"]
-  >(tilesets.IDs.Obstacles.Animal.COW)
   // The currently selected box and tool.
   const [toolbox, setToolbox] =
     useState<Phaser.Types.Scenes.Create.Toolbox.Any>({
       box: "road",
-      tool: roadTool,
+      tool: "add",
     })
-
-  // A helper function to set the tool for a specific box.
-  const setTool = <Box extends (typeof toolbox)["box"]>(box: Box) => {
-    return (tool: Phaser.Types.Scenes.Create.Toolbox.Any["tool"]) =>
-      // @ts-expect-error tool will be in box
-      setToolbox({ box, tool })
-  }
-
-  // If the toolbox changes, mirror that change in the individual tool states.
-  useEffect(() => {
-    if (toolbox.box === "road") setRoadTool(toolbox.tool)
-    else if (toolbox.box === "endpoints") setEndpointsTool(toolbox.tool)
-    else if (toolbox.box === "scenery") setSceneryTool(toolbox.tool)
-    else if (toolbox.box === "obstacles") setObstaclesTool(toolbox.tool)
-  }, [toolbox])
 
   // If react changes the toolbox, mirror that change into Phaser so that the
   // Phaser scene can react to it.
@@ -86,128 +39,47 @@ const Controls: FC<ControlsProps> = () => {
       })
   }, [phaserGame])
 
-  const makeSelectableButtonItemProps = (
-    tb: typeof toolbox,
-  ): Pick<
-    miniDrawers.ButtonItemProps,
-    "id" | "isDrawerOpen" | "selected" | "onClick"
-  > => ({
-    id: tb.box,
-    isDrawerOpen,
-    selected: toolbox.box === tb.box,
-    onClick: () => setToolbox(tb),
-  })
-
-  const toolboxControls = {
-    road: <RoadControls tool={roadTool} setTool={setTool("road")} />,
-    endpoints: (
-      <EndpointsControls tool={endpointsTool} setTool={setTool("endpoints")} />
-    ),
-    scenery: (
-      <SceneryControls tool={sceneryTool} setTool={setTool("scenery")} />
-    ),
-    obstacles: (
-      <ObstaclesControls tool={obstaclesTool} setTool={setTool("obstacles")} />
-    ),
-  }[toolbox.box]
-
   return (
     <>
-      {activeSceneKeys.includes("Create.LEVEL") && (
-        <>
-          <ZoomControls />
-          {toolboxControls}
-        </>
-      )}
-      <miniDrawers.MiniDrawer
+      {activeSceneKeys.includes("Create.LEVEL") && <ZoomControls />}
+      <MiniDrawer
         open={isDrawerOpen}
         onToggle={() => {
           setIsDrawerOpen(!isDrawerOpen)
         }}
       >
-        <miniDrawers.ButtonItem
-          {...makeSelectableButtonItemProps({ box: "road", tool: roadTool })}
-          text="Road"
-          icon={<EditRoadIcon />}
+        <items.RoadButton
+          isDrawerOpen={isDrawerOpen}
+          tool={toolbox.box === "road" ? toolbox.tool : undefined}
+          setTool={tool => setToolbox({ box: "road", tool })}
         />
-        <miniDrawers.ButtonItem
-          {...makeSelectableButtonItemProps({
-            box: "endpoints",
-            tool: endpointsTool,
-          })}
-          text="Start & End Points"
-          icon={<HomeIcon />}
+        <items.EndpointsButton
+          isDrawerOpen={isDrawerOpen}
+          tool={toolbox.box === "endpoints" ? toolbox.tool : undefined}
+          setTool={tool => setToolbox({ box: "endpoints", tool })}
         />
-        <miniDrawers.ButtonItem
-          {...makeSelectableButtonItemProps({
-            box: "obstacles",
-            tool: obstaclesTool,
-          })}
-          text="Obstacles"
-          icon={<TrafficIcon />}
+        <items.ObstaclesButton
+          isDrawerOpen={isDrawerOpen}
+          tool={toolbox.box === "obstacles" ? toolbox.tool : undefined}
+          setTool={tool => setToolbox({ box: "obstacles", tool })}
         />
-        <miniDrawers.ButtonItem
-          {...makeSelectableButtonItemProps({
-            box: "scenery",
-            tool: sceneryTool,
-          })}
-          text="Scenery"
-          icon={<ParkIcon />}
+        <items.SceneryButton
+          isDrawerOpen={isDrawerOpen}
+          tool={toolbox.box === "scenery" ? toolbox.tool : undefined}
+          setTool={tool => setToolbox({ box: "scenery", tool })}
         />
         <Divider />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Character"
-          icon={<LocalShippingIcon />}
-        />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Code"
-          icon={<ExtensionIcon />}
-        />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Random"
-          icon={<CasinoIcon />}
-        />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Description"
-          icon={<DescriptionIcon />}
-        />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Hint"
-          icon={<LightbulbIcon />}
-        />
+        <items.CharacterButton isDrawerOpen={isDrawerOpen} />
+        <items.CodeButton isDrawerOpen={isDrawerOpen} />
+        <items.DescriptionButton isDrawerOpen={isDrawerOpen} />
         <Divider />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Load"
-          icon={<CloudDownloadIcon />}
-        />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Save"
-          icon={<SaveIcon />}
-        />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Share"
-          icon={<PeopleIcon />}
-        />
+        <items.LoadButton isDrawerOpen={isDrawerOpen} />
+        <items.SaveButton isDrawerOpen={isDrawerOpen} />
+        <items.ShareButton isDrawerOpen={isDrawerOpen} />
         <Divider />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Help"
-          icon={<HelpIcon />}
-        />
-        <miniDrawers.ButtonItem
-          isDrawerOpen={isDrawerOpen}
-          text="Quit"
-          icon={<ExitToAppIcon />}
-        />
-      </miniDrawers.MiniDrawer>
+        <items.HelpButton isDrawerOpen={isDrawerOpen} />
+        <items.QuitButton isDrawerOpen={isDrawerOpen} />
+      </MiniDrawer>
     </>
   )
 }
